requirements.txt
fastapi[standard]==0.135.3
uvicorn[standard]==0.43.0
pydantic-settings==2.13.1
llama-index-core==0.14.20
llama-index-llms-openai==0.7.5
llama-index-embeddings-openai==0.6.0
llama-index-vector-stores-qdrant==0.10.0
qdrant-client==1.17.1
pdfplumber==0.11.9
httpx==0.28.1
beautifulsoup4==4.14.3
aiosqlite==0.22.1
tiktoken==0.12.0
langfuse==4.0.6
structlog==25.5.0
litellm==1.83.3
slowapi==0.1.9
fastembed==0.6.0
pytesseract==0.3.13
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.1.0
